Added ability to read Cargo.tom dependencies
Hidden behind a feature flag `make_md` README can be mostly made on
build.
The build.rs read dependencies from Cargo.toml ignore predefined
dependencies that are not being benchmarked and lookup data for those
that are. This in combination with the github script relieves any
contributor or the author of needing to edit the readme.